:art: enable NIN to be used from chainer2
