+++ neural_art/models/__init__.py
@@ -1,1 +1,2 @@
 from .vgg import *
+from .nin import *
